refactor: useFormData를 공통 hook으로 쓸 수 있게 리팩토링
